Replace error stubbing with capturing values in spy.
When calling load, we pass in a completion handler that will capture whatever error comes back from HTTPClient. We will store this completion handler in the HTTPClient. *When* we call this completion handler in our test, we will pass in whatever error that we want the HTTPClient to complete with, and that will waterfall back to the capturing. We will finally assert that what we capture matches what we passed in.